+++ src/index.ts
@@ -1,6 +1,9 @@
 export { NOW } from "./Now";
 export { Group } from "./Group";
 export { Easing } from "./Easing";
+export type { EasingFunction } from "./Easing";
 export { Interpolation } from "./Interpolation";
+export type { InterpolationFunction, ARGB, AHSV, AHCL } from "./Interpolation";
 export { Tween } from "./Tween";
+export type { RecursivePartial } from "./Tween";
 export { VERSION } from "./Version";