+++ test/TweenEvents.test.ts
@@ -63,15 +63,19 @@ test("Tween calls onRepeat every times it repeats", () => {
 	g.update(50);
 
 	expect(fn).toHaveBeenCalledTimes(1);
+	expect(fn).toHaveBeenLastCalledWith(expect.anything(), 1, expect.anything());
 
 	g.update(100);
 	expect(fn).toHaveBeenCalledTimes(2);
+	expect(fn).toHaveBeenLastCalledWith(expect.anything(), 2, expect.anything());
 
 	g.update(100);
 	expect(fn).toHaveBeenCalledTimes(3);
+	expect(fn).toHaveBeenLastCalledWith(expect.anything(), 3, expect.anything());
 
 	g.update(100);
 	expect(fn).toHaveBeenCalledTimes(3);
+	expect(fn).toHaveBeenLastCalledWith(expect.anything(), 3, expect.anything());
 });
 
 test("Tween calls onStop only when manually stopped", () => {
@@ -103,6 +107,20 @@ test("Tween calls onStart when the first update is called", () => {
 	new Tween({ a: 0 }, g).to({ a: 1 }, 100).start().onStart(fn1);
 	expect(fn1).not.toHaveBeenCalled(); // not yet
 	g.update(1);
+	expect(fn1).toHaveBeenCalled(); // now!
+	g.update(1);
+	expect(fn1).toHaveBeenCalledTimes(1); // but only once tho
+});
+
+test("Tween calls onAfterDelay after the delay timer runs out", () => {
+	const g = new Group();
+
+	const fn1 = jest.fn();
+	new Tween({ a: 0 }, g).to({ a: 1 }, 100).delay(50).start().onAfterDelay(fn1);
+	expect(fn1).not.toHaveBeenCalled(); // not yet
+	g.update(25);
+	expect(fn1).not.toHaveBeenCalled(); // not yet...
+	g.update(25);
 	expect(fn1).toHaveBeenCalled(); // now!
 	g.update(1);
 	expect(fn1).toHaveBeenCalledTimes(1); // but only once tho